moved client and server to http directory
