--- crates/rotala/src/client/mod.rs
@@ -1,5 +1,0 @@
-//! Client implements an Exchange API over the network so that the interface called is the same
-//! on the network as locally.
-//!
-//! Uses reqwest to send requests to server.
-pub mod uist;
