+++ crates/rotala/src/lib.rs
@@ -73,10 +73,9 @@
 //!
 //! Long-term:
 //! - Add orderbook with L2 data, this is going to require L2 sources
-pub mod client;
 pub mod clock;
 pub mod exchange;
+pub mod http;
 pub mod input;
 pub mod orderbook;
-pub mod server;
 pub mod source;
