--- crates/rotala/src/server/mod.rs
@@ -1,2 +1,0 @@
-//! Server that implements an exchange. All servers should use JSON on request and response.
-pub mod uist;
